+++ lib/packages/wikibase/data-model/src/Entity/Item.php
@@ -9,6 +9,7 @@ use Wikibase\DataModel\SiteLinkList;
 use Wikibase\DataModel\Statement\StatementList;
 use Wikibase\DataModel\Statement\StatementListHolder;
 use Wikibase\DataModel\Term\Fingerprint;
+use Wikibase\DataModel\Term\FingerprintHolder;
 
 /**
  * Represents a single Wikibase item.
@@ -19,7 +20,7 @@ use Wikibase\DataModel\Term\Fingerprint;
  * @licence GNU GPL v2+
  * @author Jeroen De Dauw < jeroendedauw@gmail.com >
  */
-class Item extends Entity implements StatementListHolder {
+class Item implements EntityDocument, FingerprintHolder, StatementListHolder {
 
 	const ENTITY_TYPE = 'item';