+++ lib/packages/wikibase/data-model/src/Entity/Property.php
@@ -6,6 +6,7 @@ use InvalidArgumentException;
 use Wikibase\DataModel\Statement\StatementList;
 use Wikibase\DataModel\Statement\StatementListHolder;
 use Wikibase\DataModel\Term\Fingerprint;
+use Wikibase\DataModel\Term\FingerprintHolder;
 
 /**
  * Represents a single Wikibase property.
@@ -16,7 +17,7 @@ use Wikibase\DataModel\Term\Fingerprint;
  * @licence GNU GPL v2+
  * @author Jeroen De Dauw < jeroendedauw@gmail.com >
  */
-class Property extends Entity implements StatementListHolder {
+class Property implements EntityDocument, FingerprintHolder, StatementListHolder {
 
 	const ENTITY_TYPE = 'property';